examples/v1/basic-hello/requirements.txt
# Programming Model v1 example dependencies
azure-functions
azure-functions-doctor
